Various tests for the new feature: filename bruteforce. All tests pass with the latest moth.
