+++ plugins/tests/crawl/dir_file_bruter/test_files_small.db
@@ -1,0 +1,7 @@
+donotexist.png
+donotexist.bmp
+donotexist.txt
+iamhidden.txt
+donotexist.db
+donotexist.sqlite3
+db.sqlite3
